requirements.txt
bs4==0.0.1
pymongo==3.11.2
pyyaml==5.3.1
requests==2.25.1
selenium==3.141.0